Move simd module to a new simd-array crate.
Having the SIMD array code in a separate crate makes our life a bit
easier, since e.g. benchmarks are not supported in a `cdylib` project.
