+++ simd-array/src/array.rs
@@ -1,19 +1,19 @@
 #[cfg(target_arch = "aarch64")]
 use std::arch::is_aarch64_feature_detected;
 
-use crate::simd::activation::Activation;
-use crate::simd::distribution::Distribution;
 use num_traits::Float;
 
+use crate::activation::Activation;
+use crate::distribution::Distribution;
 #[cfg(target_arch = "x86_64")]
-use crate::simd::vector::avx::AVXVector32;
+use crate::vector::avx::AVXVector32;
 #[cfg(target_arch = "x86_64")]
-use crate::simd::vector::avx::AVXVector64;
+use crate::vector::avx::AVXVector64;
 #[cfg(target_arch = "aarch64")]
-use crate::simd::vector::neon::NeonVector32;
+use crate::vector::neon::NeonVector32;
 #[cfg(target_arch = "aarch64")]
-use crate::simd::vector::neon::NeonVector64;
-use crate::simd::vector::{ScalarVector32, ScalarVector64, SimdVector};
+use crate::vector::neon::NeonVector64;
+use crate::vector::{ScalarVector32, ScalarVector64, SimdVector};
 
 #[cfg(target_arch = "aarch64")]
 pub fn platform_arrays() -> (Box<dyn Array<Scalar = f32>>, Box<dyn Array<Scalar = f64>>) {
